requirements.txt

fastapi
uvicorn
azure-functions

